Fix readVarint35(): accumulate uint64
Closes clarisma/geodesk-gol#43

+++ include/clarisma/util/varint.h
@@ -18,16 +18,16 @@ inline uint64_t readVarint35(const uint8_t*& p)
 	val = b & 0x7f;
 	if ((b & 0x80) == 0) return val;
 	b = *p++;
-	val |= static_cast<uint32_t>(b & 0x7f) << 7;
+	val |= static_cast<uint64_t>(b & 0x7f) << 7;
 	if ((b & 0x80) == 0) return val;
 	b = *p++;
-	val |= static_cast<uint32_t>(b & 0x7f) << 14;
+	val |= static_cast<uint64_t>(b & 0x7f) << 14;
 	if ((b & 0x80) == 0) return val;
 	b = *p++;
-	val |= static_cast<uint32_t>(b & 0x7f) << 21;
+	val |= static_cast<uint64_t>(b & 0x7f) << 21;
 	if ((b & 0x80) == 0) return val;
 	b = *p++;
-	val |= static_cast<uint32_t>(b & 0x7f) << 28;
+	val |= static_cast<uint64_t>(b & 0x7f) << 28;
 	assert((b & 0x80) == 0);
 	return val;
 }
